requirements.txt
opencv-python
pyautogui
numpy
Pillow
requests
beautifulsoup4
tkinter
